Control backend ros bridge service implemented with publish method, delete the unnecessary folder, unsubscribe unnecessary node

+++ control-backend/src/main/java/com/campusride/model/Scooter.java
@@ -7,7 +7,7 @@ public class Scooter {
     private double y;
     private String status;
     private int battery;
-    private int speed;
+    private float speed;
 
     // Getters and setters
     public int getId() { return id; }
@@ -28,6 +28,6 @@ public class Scooter {
     public int getBattery() { return battery; }
     public void setBattery(int battery) { this.battery = battery; }
 
-    public int getSpeed() { return speed; }
-    public void setSpeed(int speed) { this.speed = speed; }
+    public float getSpeed() { return speed; }
+    public void setSpeed(float speed) { this.speed = speed; }
 }
